+++ src/plugin.js
@@ -1,7 +1,9 @@
 import MagicString from 'magic-string'
+import { formatters } from './formatters'
 
 const defaultOptions = {
   sourcemap: true,
+  formatter: 'identity',
 }
 export default function banner2(resolveBanner, userOptions) {
   const opts = { ...defaultOptions, ...(userOptions || {}) }
@@ -11,10 +13,14 @@ export default function banner2(resolveBanner, userOptions) {
     async renderChunk(code, chunk, options) {
       const banner = await resolveBanner(chunk, options)
       if (!banner) return { code, map: null }
-      if (!opts.sourcemap) return banner + code
+
+      const formatterByOption = formatters[opts.formatter]
+      const formattedBanner = formatterByOption(String(banner))
+
+      if (!opts.sourcemap) return formattedBanner + code
 
       const magicString = new MagicString(code)
-      magicString.prepend(String(banner))
+      magicString.prepend(formattedBanner)
 
       return {
         code: magicString.toString(),